Hooks for login/register forms instead of 1 4 both
